Introduce the sparkle executable.
Only one subcommand is available for the moment:

    $ sparkle package <some-exe-in-the-PATH>

This command replaces the `mvn package` step of old. Since the .jar
for sparkle now ships as a data-file, all that's left to do for an app
is to decompress the jar, create an app.zip file containing all
dynlibs, and then injecting that in the jar. We don't need fancy pants
Java build systems for that, be it Gradle or Maven. We can do it
ourselves, in Haskell, and that's what the sparkle command is.

With the advent of this command, copy-exe.sh is obsolete.