Implement jarify as a set of Bazel rules, rather than Haskell code.
This is essentially a rewrite of the whole project. We're changing the
name of the project while we're at it, since the tool now does a lot
less. Where the tool used to inject binaries into JAR's, now we
delegate this to existing Bazel functionality. Instead, the tool
focuses solely on computing closures given sets of shared objects.