Don't depend on cc_haskell_import to rename binaries.
cc_haskell_import was intended to solve a different problem. Where
we only needed renaming the input files, the rule would add new files
to the output complicating the processing downstream.

Now library_closure is easier to use as the caller was the one
needing to use cc_haskell_import. This is no longer necessary.